% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroTelo_dist.R
\name{centroTelo_dist}
\alias{centroTelo_dist}
\title{Compute proximity with closest telomere and centromere}
\usage{
centroTelo_dist(chr, position)
}
\arguments{
\item{chr}{the chromosome number}

\item{position}{the locus position}
}
\value{
telomere.bool: boolean value equal to TRUE if the position is inside a telomere region

telomere.prox: proximity value to the closest telomere

centromere.bool: boolean value equal to TRUE if the position is inside a centromere region

centromere.prox: proximity value to the closest centromere
}
\description{
\code{centroTelo_dist}
}
\examples{
#get centro/telo -mere related features for locus chr1:5000000
centroTelo_dist(chr=1,position=5000000)
}
