% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dDMPs_dist.R
\name{dDMPs_dist}
\alias{dDMPs_dist}
\title{Compute proximity with closest developmental brain methylation spots (regions or positions)}
\usage{
dDMPs_dist(chr, position, verbose = T)
}
\arguments{
\item{chr}{the chromosome number}

\item{position}{the locus position}
}
\value{
dDMP-pos-bool: boolean value equal to TRUE if the position is a known dDMP position

dDMP-pos-prox: proximity value to the closest dDMP position

dDMP-reg-bool: boolean value equal to TRUE if the position is a known dDMP region

dDMP-reg-prox: proximity value to the closest dDMP region
}
\description{
\code{dDMPs_dist}
}
\examples{
#get dDMPs related features for locus chr1:5000000
dDMPs_dist(chr=1,position=5000000)

}
