% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{feature_extraction}
\alias{feature_extraction}
\title{Compute the feature representation for one or more genomic positions}
\usage{
feature_extraction(chr = 1, position = 5e+06)
}
\arguments{
\item{chr}{chromosome number(s)}

\item{position}{genomic position(s)}
}
\value{
a named matrix for feature space representation (~ 360 features), each row corresponds to one input position
}
\description{
\code{feature_extraction}
}
\examples{
#get features for locus chr1:5000000
feature_extraction(chr=c(1,2),position=c(5000000,10000000))

}
