% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams_count.R
\name{get_ngrams_count}
\alias{get_ngrams_count}
\title{Compute the frequency for each 1,2,3,4-nucleotide in a +/-500bp neighbourhood around a locus}
\usage{
get_ngrams_count(chr = "chr1", position = 5e+06)
}
\arguments{
\item{chr}{the chromosome number (including 'chr' prefix)}

\item{position}{the locus position, center of 1kbp window}
}
\value{
a named vector of length 340 (4 + 16 + 64 + 256) with ngram frequencies
}
\description{
\code{ngrams}
}
\examples{
#get ngrams count in chr1-4999500:5000500 interval
get_ngrams_count(chr=1,position=5000000)
}
