% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClosestGene.R
\name{getClosestGene}
\alias{getClosestGene}
\title{Compute proximity with closest tissue and non-tissue genes}
\usage{
getClosestGene(chr, position)
}
\arguments{
\item{chr}{the chromosome number}

\item{position}{the locus position}
}
\value{
tissue.gene.bool: boolean value equal to TRUE if the position is inside a known tissue gene

tissue.gene.prox: proximity value to the closest tissue gene

bgd.gene.bool: boolean value equal to TRUE if the position is inside a non-tissue gene

bgd.gene.prox: proximity value to the closest non-tissue gene
}
\description{
\code{getClosestGene}
}
\examples{
#get genes related features for locus chr1:5000000
getClosestGene(chr=1,position=5000000)

}
