% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getIDs}
\alias{getIDs}
\title{Get access to pubmed ID given query on keywords in title/abstract/etc}
\usage{
getIDs(term)
}
\arguments{
\item{term}{a search term (character vector of length 1)}

\item{position}{the locus position}
}
\value{
out: a character vector of PMIDs
}
\description{
\code{getIDs}
}
\examples{
#get PubMed IDs for a query
getIDs("heart[Title]")
}
