% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTEx_eQTL.R
\name{GTEx_eQTL}
\alias{GTEx_eQTL}
\title{Compute proximity with closest tissue and non-tissue GTEx eQTLs}
\usage{
GTEx_eQTL(chr, position)
}
\arguments{
\item{chr}{the chromosome number}

\item{position}{the locus position}
}
\value{
bool.tissue: boolean value equal to TRUE if the position is a known tissue GTEx eQTL

prox.tissue: proximity value to the closest tissue eQTL

bool.other: boolean value equal to TRUE if the position is a known non-tissue GTEx eQTL

prox.other: proximity value to the closest non-tissue eQTL
}
\description{
\code{GTEx_eQTL}
}
\examples{
#get GTEx eQTL related features for locus chr1:5000000
GTEx_eQTL(chr=1,position=5000000)

}
