% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RME_methyl.R
\name{RME_methyl}
\alias{RME_methyl}
\title{Compute proximity with closest tissue and non-tissue RoadMap Epigenomics(RME) database}
\usage{
RME_methyl(chr, position, verbose = T)
}
\arguments{
\item{chr}{the chromosome number}

\item{position}{the locus position}
}
\value{
RME.bool: boolean value equal to TRUE if the position is inside a RME methylated region

RME.prox: proximity value to the closest RME methylated region

RME.tissue: mean methylation value across all tissues of interest  (only if belongs to a methylated region)

RME.others: mean methylation value across all other tissues (only if belongs to a methylated region)
}
\description{
\code{RME_methyl}
}
\examples{
#get tissue-specific methylation related features for locus chr1:5000000
RME_methyl(chr=1,position=5000000)

}
